% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{example_snp_data}
\alias{example_snp_data}
\title{Simulated example dataset of SNPs obtained during a GWAS}
\format{
A matrix with 200 rows and 300 variables with integer values being either 0, 1 or 2.
}
\usage{
example_snp_data
}
\description{
A dataset containing the numeric interpretation of 500 SNPs from 200 simulated humans. Both
dimensions have a number-based names attribute.
}
\keyword{datasets}
