% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagecoxph.R
\name{firststagecoxph.multicore}
\alias{firststagecoxph.multicore}
\title{Perform the first stage of a GWAS in parallel}
\usage{
firststagecoxph.multicore(
  survival.dataset,
  covariate.matrix,
  progress = 50,
  max.coef = 5,
  max.batchsize = 1000,
  updatefile = "",
  snps.are.named = FALSE
)
}
\arguments{
\item{survival.dataset}{the outcomes}

\item{covariate.matrix}{the covariates}

\item{progress}{how often progress must be reported}

\item{max.coef}{what the maximum coefficients may be when fitting}

\item{max.batchsize}{what the maximum batchsize must be.}

\item{updatefile}{path to file to replace terminal}

\item{snps.are.named}{wheter or not the covariates have names}
}
\value{
the p-values of the first stage (possibly in random order) with proper names() attribute.
As a consequence of possible empty batches, there may be trailing NA's. Ironically however,
these are  not guaranteed to be trailing, due to the fact that batches can finish
in any order. They are identified by their names attribute being empty: "".
}
\description{
Perform the first stage of a GWAS in parallel
}
\details{
The covariates are split into a number of batches. The amount of covariates in every batch
is optimal, i.e. the load is distributed equally over all batches, the number
of covariates in each batch does not exceed max.batchsize, and the number of batches is a
multiple of the number of worker-processes.
These batches are then analysed in parallel by the worker-processes that return the
corresponding p-values of the marginal associations of the covariates. These p-values have
their names attribute set to the index of the covariate, so that the processes do not need
to finish in order. If they do not, the order of the returned p-values will not match up
with the order in which the covariates were provided, so this problem is avoided.
}
