% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagecoxph.R
\name{optimal.batch.configuration}
\alias{optimal.batch.configuration}
\title{Calculates the optimal batch configuration}
\usage{
optimal.batch.configuration(no.covariates, max.batchsize, no.workers = 1)
}
\arguments{
\item{no.covariates}{number of covariates to distribute}

\item{max.batchsize}{upper limit on amount covariates in each batch}

\item{no.workers}{number of workers}
}
\value{
a list of size 3, containing the number of covariates in each batch, the number of batches
(including possible empty ones at the end, this way it always is a multiple of no.workers), and a vector of batchsizes for the last iteration (excluding trailing zeros)
}
\description{
Calculates the optimal batch configuration
}
\details{
configures this in such a way that the number of covariates does not exceed the
allowed maximum, while having each batch filled equally (not necessarily full). The last batch
may be not fully filled. Made for the first stage, but easily adaptible to the second stage.
}
