% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagecoxph.R
\name{print.twostageGWAS}
\alias{print.twostageGWAS}
\title{Print method for twostage object}
\usage{
\method{print}{twostageGWAS}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{passed}
}
\value{
the twostage object
}
\description{
Print method for twostage object
}
\examples{
print("example")
}
