% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagecoxph.R
\name{singlecore.twostagecoxph}
\alias{singlecore.twostagecoxph}
\title{Two stage method using single core}
\usage{
singlecore.twostagecoxph(
  survival.dataset,
  covariate.matrix,
  first.stage.threshold
)
}
\arguments{
\item{survival.dataset}{the dataset}

\item{covariate.matrix}{the covariates}

\item{first.stage.threshold}{the fst}
}
\value{
a sparseMatrix object
}
\description{
Two stage method using single core
}
\examples{
print("make exmple")
}
