% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{example_survival_data}
\alias{example_survival_data}
\title{Simulated example dataset for a survival GWAS}
\format{
A Surv-class object with 200 rows and 2 variables:
\describe{
\item{time}{the time until the event took place}
\item{status}{indicator whether or not the subject was censored. 1 if the event was observed.}
}
}
\usage{
example_survival_data
}
\description{
A dataset containing time-to-event outcomes for 200 subjects, simulated from a proportional
hazards model with one interaction increasing the hazard. The censoring times have a uniform
distribution.
}
\keyword{datasets}
