% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagecoxph.R
\name{print.twostageGWAS}
\alias{print.twostageGWAS}
\title{Print method for twostage object}
\usage{
\method{print}{twostageGWAS}(x, ...)
}
\arguments{
\item{x}{object of class twostageGWAS}

\item{...}{optional arguments passed on to cat and print.default functions.}
}
\value{
The twostageGWAS object invisibly
}
\description{
Print method for twostage object
}
\examples{
survival.dataset <- survival::Surv(c(5,5,3,3,2,2,2,1,1,1),
                                   c(0,0,1,1,1,1,1,1,1,1))
covariate.matrix <- matrix(c(2,2,1,
                             2,2,1,
                             1,2,1,
                             2,1,1,
                             1,1,1,
                             1,1,1,
                             1,0,0,
                             0,1,0,
                             1,0,0,
                             0,0,0),
                           nrow = 10, ncol = 3, byrow = TRUE)
twostagecoxph(survival.dataset, covariate.matrix, progress = 0)
}
