% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagecoxph.R
\name{twostagecoxph}
\alias{twostagecoxph}
\title{(WIP) Perform a two stage analysis on a survival dataset to detect epistasis}
\usage{
twostagecoxph(
  survival.dataset,
  covariate.matrix,
  first.stage.threshold = 0.05,
  multiple.hypotheses.correction = "bonferroni",
  multicore = FALSE,
  report.lowest.amount = 5,
  return.raw = FALSE,
  progress = 1000,
  max.coef = 5
)
}
\arguments{
\item{survival.dataset}{The survival dataset describing the outcome.}

\item{covariate.matrix}{The nxp-matrix of covariates of the p covariates of the n patients.}

\item{first.stage.threshold}{numeric scalar denoting the threshold for the first stage. If a covariate
has a p-value lower than this threshold, it will be passed on to the second stage.}

\item{multiple.hypotheses.correction}{Correction method, a character string. Passed to stats::p.adjust}

\item{multicore}{Currently unused}

\item{report.lowest.amount}{integer scalar denoting how many of the most significant interactions
the function should report}

\item{return.raw}{logical, default FALSE; whether or not the output should contain the raw p-values or
the multiple hypotheses corrected p-values.}

\item{progress}{numeric, default 1000;  how many iterations should pass silently until an update is given
about runtime and progress until completion of stages. Set to 0 for no output.}

\item{max.coef}{numeric, default 5; maximum value for all coefficients in the fitted models. If any
are larger than this (in absolute value), then the model rejected and ignored in
further analysis.}
}
\value{
Hopefully a nicely structured list of some kind?
}
\description{
Work in progress. Only the most basic functionality works.
}
\note{
Parallel processing is currently not supported
}
\examples{
survival.dataset <- survival::Surv(c(5,5,3,3,2,2,2,1,1,1),
                                   c(0,0,1,1,1,1,1,1,1,1))
covariate.matrix <- matrix(c(2,2,1,
                             2,2,1,
                             1,2,1,
                             2,1,1,
                             1,1,1,
                             1,1,1,
                             1,0,0,
                             0,1,0,
                             1,0,0,
                             0,0,0),
                           nrow = 10, ncol = 3, byrow = TRUE)
twostagecoxph(survival.dataset, covariate.matrix, progress = 0)

## Not run:
#load("data/example_survival_data.rda")
#load("data/example_snp_data.rda")
#str(example_survival_data)
#str(example_snp_data)
#foo <- twostagecoxph(example_survival_data, example_snp_data, first.stage.threshold = 1e-5)
## End(Not run)
}
\seealso{
print.twostageGWAS()
}
