% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagecoxph.R
\name{twostagecoxph}
\alias{twostagecoxph}
\title{(WIP) Perform a two stage analysis on a survival dataset to detect epistasis}
\usage{
twostagecoxph(
  survival.dataset,
  covariate.matrix,
  first.stage.threshold = 0.05,
  multiple.hypotheses.correction = "bonferroni",
  multicore = FALSE,
  updatefile = "",
  control = twostagecoxph.control()
)
}
\arguments{
\item{survival.dataset}{The survival dataset describing the outcome.}

\item{covariate.matrix}{The nxp-matrix of covariates of the p covariates of the n patients.}

\item{first.stage.threshold}{numeric scalar denoting the threshold for the first stage. If a covariate
has a p-value lower than this threshold, it will be passed on to the second stage.}

\item{multiple.hypotheses.correction}{Correction method, a character string. Passed to stats::p.adjust}

\item{multicore}{logical, default FALSE; whether or not the function should use multiple cores
in its calculations. See Details.}

\item{updatefile}{path to a text file where updates may be written. Necessary for parallel
computations, since the connection to the terminal will be lost. This
file will serve as a stand in for the terminal.}

\item{control}{Object of class twostagecoxph.control specifying various options for performance
of the two stage method. Default is twostagecoxph.control()}
}
\value{
A twostageGWAS object which is a list of 7 entries:
\item{most.significant.results}{A list describing the most significant results found. The
number of results reported is specified by the control parameter, default 5. The
list contains 3 items:
\describe{
\item{\code{interacting.snps}}{the names of the interaction, in "name_snp_1 x name_snp_2" format.
In case the covariates in \code{covariate.matrix} have a names attribute, these names are used.
Otherwise the index of the covariates within the matrix is used.}
\item{\code{p.value.epistasis}}{the corresponding p-values of the interactions. Unless \code{return.raw = TRUE}
is specified in the control parameter, these p-values will be corrected for the multiple hypotheses
tested with the method specified by the \code{multiple.hypotheses.correction} parameter.}
\item{\code{duplicate.interactions}}{the list of duplicate interactions found, corresponding to
the interactions specified in this list. An interaction is said to be duplicate if the corresponding
p-value is the same up until 7 significant figures. }
}}
\item{p.value.matrix}{A \code{sparseMatrix} from the package "Matrix", specifying the resulting
upper triangular p-value matrix obtained from the second stage.  Unless \code{return.raw = TRUE}
is specified in the control parameter, these p-values will be corrected for the multiple hypotheses
tested with the method specified by the \code{multiple.hypotheses.correction} parameter.}
\item{marginal.significant}{A vector of named integers, specifying the indices of covariates which
were found to be marginally significant in the first stage. }
\item{first.stage}{A vector specifying the p-values found in the first stage. These p-values
are \emph{not} corrected for the multiple tested hypotheses. }
\item{fst}{The threshold for significance used in the first stage as specified in \code{first.stage.threshold}}
\item{runtime}{The total runtime of the function in seconds.}
\item{call}{The matched call.}
}
\description{
Work in progress. Only basic functionality works.
}
\note{
Parallel processing requires a properly registered parallel back-end, such as one obtained
from "doParallel::registerDoParallel(2)" to be used by \code{foreach} and the \code{\%dopar\%} binary operator.
}
\examples{
survival.dataset <- survival::Surv(c(5,5,3,3,2,2,2,1,1,1),
                                   c(0,0,1,1,1,1,1,1,1,1))
covariate.matrix <- matrix(c(2,2,1,
                             2,2,1,
                             1,2,1,
                             2,1,1,
                             1,1,1,
                             1,1,1,
                             1,0,0,
                             0,1,0,
                             1,0,0,
                             0,0,0),
                           nrow = 10, ncol = 3, byrow = TRUE)
twostagecoxph(survival.dataset, covariate.matrix, control = twostagecoxph.control(progress = 0))

str(example_survival_data)
str(example_snp_data)

print(foo <- twostagecoxph(example_survival_data, example_snp_data[,1:300],
                            first.stage.threshold = 1e-5))
print(bar <- twostagecoxph(example_survival_data, example_snp_data[,1:300],
                            first.stage.threshold = 1e-4))
#[,1:300] subsetting is added to speed up the example. Try removing it! :)

# As we can see, foo and bar have different results. A lower FST generally gives more power, but it
# it risks the possibility to be too strict and consequently *decreasing* power.
}
\seealso{
\code{\link{foreach}}, \code{\link{print.twostageGWAS}}, \code{\link{twostagecoxph.control}}
}
