% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostagecoxph.R
\name{twostagecoxph}
\alias{twostagecoxph}
\title{Perform a two stage analysis on a survival dataset to detect epistasis}
\usage{
twostagecoxph(
  survival.dataset,
  covariate.matrix,
  first.stage.threshold = 0.05,
  multiple.hypotheses.correction = "bonferroni",
  multicore = FALSE,
  report.lowest.amount = 5,
  return.raw = FALSE
)
}
\arguments{
\item{survival.dataset}{The survival dataset describing the outcome.}

\item{covariate.matrix}{The nxp-matrix of covariates of the p covariates of the n patients.}

\item{first.stage.threshold}{numeric scalar denoting the threshold for the first stage. If a covariate
has a p-value lower than this threshold, it will be passed on to the second stage.}

\item{multiple.hypotheses.correction}{Correction method, a character string. Passed to stats::p.adjust}

\item{multicore}{Currently unused}

\item{report.lowest.amount}{integer scalar denoting how many of the most significant interactions
the function should report}

\item{return.raw}{logical, default FALSE; whether or not the output should contain the raw p-values or
the multiple hypotheses corrected p-values.}
}
\value{
Hopefully a nicely structured list of some kind?
}
\description{
Perform a two stage analysis on a survival dataset to detect epistasis
}
\note{
Parallel processing is currently not supported
}
\examples{
survival.dataset <- survival::Surv(c(5,5,3,3,2,2,2,1,1,1),
                                   c(0,0,1,1,1,1,1,1,1,1))
covariate.matrix <- matrix(c(2,2,1,
                             2,2,1,
                             1,2,1,
                             2,1,1,
                             1,1,1,
                             1,1,1,
                             1,0,0,
                             0,1,0,
                             1,0,0,
                             0,0,0),
                           nrow = 10, ncol = 3, byrow = TRUE)
twostagecoxph(survival.dataset, covariate.matrix)
}
